Flow around cylinder example in 2D running and performing drag/lift post processing. Reynolds number is controlled by new parameter databse, as had been the case for 3D already.

+++ src/Examples/Example_NSE3D.C
@@ -204,7 +204,7 @@ Example_NSE3D::Example_NSE3D(const ParameterDatabase& user_input_parameter_db)
       // this is just an example to illustrate how a value given by
       // the user in the input file can be used in the example
       this->example_database["variable_parameter1"] = 500;
-      flow_around_cylinder_stat::VARIABLE_PARAMETER1 = this->example_database["variable_parameter1"];
+      flow_around_cylinder_stat::VARIABLE_PARAMETER1 = (size_t) this->example_database["variable_parameter1"];
 
       ExampleFile();
       break;
